requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-telegram-bot>=20.0
imaplib2>=0.57
python-decouple>=3.8
requests>=2.31.0
schedule>=1.2.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
jira>=3.5.0
gitpython>=3.1.0
openpyxl>=3.1.0
python-docx>=0.8.11
pillow>=10.0.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
fastapi>=0.104.0
uvicorn>=0.24.0
aiofiles>=23.2.1
httpx>=0.25.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
redis>=5.0.0
celery>=5.3.0
python-multipart>=0.0.6
email-validator>=2.1.0